refactor(data.js): split into a levels module
