+++ src/data.js
@@ -1,65 +1,4 @@
-/* Data Module */
-
-const TEST_LEVEL = [
-	[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,8,0,0,0,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,0,4,4,0,8,0,0,0,0,0,0,0,0],
-	[0,0,0,0,0,0,4,4,0,8,4,4,0,0,0,0,0,0,0,0,4,0],
-	[0,0,0,0,0,0,4,4,0,0,4,4,8,0,8,0,8,0,0,0,4,0],
-	[0,0,0,0,0,0,4,4,4,0,0,0,0,8,0,0,0,8,0,4,4,0],
-	[0,0,0,0,0,0,0,4,4,4,4,4,4,4,0,0,4,4,0,0,0,0],
-	[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-];
-
-export const LARGE_LEVEL = [
-	[4,4,4,7,0,0,0,0,0,0,0,0,4,0,4,0,4,4,4,7,0,4,0,4,4,4,4,4,7,0,4,0,4,0,4,4,7,0,0,0,0,0,0,0],
-	[4,7,0,0,0,0,0,0,0,0,0,0,4,0,7,0,4,4,4,4,0,4,0,7,4,4,4,4,4,0,4,0,7,0,4,4,4,0,0,0,0,0,0,0],
-	[4,0,0,0,8,0,0,4,4,0,0,0,0,4,0,0,0,0,4,4,0,0,4,0,0,0,4,4,4,0,0,4,0,0,0,4,4,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,4,4,4,4,0,0,4,4,4,4,4,0,0,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0],
-	[4,4,7,0,0,0,0,0,0,0,0,0,4,0,0,0,0,4,0,0,0,4,0,4,0,0,0,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,7,0,0,0,0,0,4,0,4,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,0,0,0,0,0],
-	[4,0,0,0,0,0,0,4,4,4,0,4,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,4,0,0,4,0,0,0,0,0,0,0,0,0,0],
-	[4,8,0,0,0,4,8,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,4,4,7,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,7,0,0,0,0,0,4,0,0,0,4,0,4,0,4,4,4,4,0,4,0,4,0,4,4,0,0,0,4,0,4,0,4,4,4,0,4,4,0,4,4,0],
-	[4,0,4,0,4,7,0,0,4,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,4,7,0,0,0,0,0,0,0,4,7,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,0,0,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,0,0,0,0,0,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-];
-
-
-export const LEVEL = [
-	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,7,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-	[4,4,4,4,4,4,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,4,4,4],
-	[4,4,4,4,4,4,4,4,7,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,4,4,4,4,4,4,7,4,0,4,4,4,4,0,4,4,4],
-	[4,4,4,4,4,4,4,0,4,7,4,4,0,0,0,8,0,0,0,0,0,4,4,4],
-	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,0,0,0,0,0,4,4,4],
-	[4,4,4,4,4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,4,4,4,4],
-	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,0,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4],
-	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4],
-];
+/* Data */
 
 export const SPRITE_LOOKUP = {
 	0: {
