+++ src/levels.js
@@ -1,0 +1,61 @@
+/* Levels */
+
+export const TEST_LEVEL = [
+	[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,8,0,0,0,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,0,4,4,0,8,0,0,0,0,0,0,0,0],
+	[0,0,0,0,0,0,4,4,0,8,4,4,0,0,0,0,0,0,0,0,4,0],
+	[0,0,0,0,0,0,4,4,0,0,4,4,8,0,8,0,8,0,0,0,4,0],
+	[0,0,0,0,0,0,4,4,4,0,0,0,0,8,0,0,0,8,0,4,4,0],
+	[0,0,0,0,0,0,0,4,4,4,4,4,4,4,0,0,4,4,0,0,0,0],
+	[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+];
+
+export const LARGE_LEVEL = [
+	[4,4,4,7,0,0,0,0,0,0,0,0,4,0,4,0,4,4,4,7,0,4,0,4,4,4,4,4,7,0,4,0,4,0,4,4,7,0,0,0,0,0,0,0],
+	[4,7,0,0,0,0,0,0,0,0,0,0,4,0,7,0,4,4,4,4,0,4,0,7,4,4,4,4,4,0,4,0,7,0,4,4,4,0,0,0,0,0,0,0],
+	[4,0,0,0,8,0,0,4,4,0,0,0,0,4,0,0,0,0,4,4,0,0,4,0,0,0,4,4,4,0,0,4,0,0,0,4,4,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,4,4,4,4,0,0,4,4,4,4,4,0,0,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0],
+	[4,4,7,0,0,0,0,0,0,0,0,0,4,0,0,0,0,4,0,0,0,4,0,4,0,0,0,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,7,0,0,0,0,0,4,0,4,0,0,0,0,0,0,0,0,0,0,7,0,0,0,0,0,4,4,0,4,0,0,0,0,0,0,0,0,0,0,0],
+	[4,0,0,0,0,0,0,4,4,4,0,4,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,4,0,0,4,0,0,0,0,0,0,0,0,0,0],
+	[4,8,0,0,0,4,8,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,4,4,7,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,7,0,0,0,0,0,4,0,0,0,4,0,4,0,4,4,4,4,0,4,0,4,0,4,4,0,0,0,4,0,4,0,4,4,4,0,4,4,0,4,4,0],
+	[4,0,4,0,4,7,0,0,4,0,0,0,0,0,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,4,7,0,0,0,0,0,0,0,4,7,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,0,0,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,0,0,0,0,0,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+];
+
+export const LEVEL = [
+	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,7,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
+	[4,4,4,4,4,4,0,8,0,0,0,0,0,0,0,0,0,0,0,0,0,4,4,4],
+	[4,4,4,4,4,4,4,4,7,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,4,4,4,4,4,4,7,4,0,4,4,4,4,0,4,4,4],
+	[4,4,4,4,4,4,4,0,4,7,4,4,0,0,0,8,0,0,0,0,0,4,4,4],
+	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,0,0,0,0,0,4,4,4],
+	[4,4,4,4,4,4,4,4,4,4,4,4,7,0,0,0,0,0,0,0,4,4,4,4],
+	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,0,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4],
+	[4,4,4,4,4,4,4,0,4,4,4,4,4,4,0,0,4,4,4,4,4,4,4,4],
+];
